requirements.txt
Flask==2.3.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
Pillow==10.0.1
requests==2.31.0
APScheduler==3.10.4
selenium==4.15.0
webdriver-manager==4.0.1
playwright==1.40.0
python-dotenv==1.0.0